+++ server/error/error_test.go
@@ -26,4 +26,50 @@ func TestError_StatusCode(t *testing.T) {
 	if e.StatusCode() != http.StatusServiceUnavailable {
 		t.Fatalf("TestError_StatusCode %v failed", e)
 	}
+
+	if !e.InternalError() {
+		t.Fatalf("TestInternalError failed")
+	}
+}
+
+func TestNewError(t *testing.T) {
+	var err error
+	err = NewError(ErrInvalidParams, "test1")
+	if err == nil {
+		t.Fatalf("TestNewError failed")
+	}
+	err = NewErrorf(ErrInvalidParams, "%s", "test2")
+	if err == nil {
+		t.Fatalf("TestNewErrorf failed")
+	}
+
+	if len(err.Error()) == 0 {
+		t.Fatalf("TestError failed")
+	}
+
+	if len(err.(*Error).Marshal()) == 0 {
+		t.Fatalf("TestMarshal failed")
+	}
+
+	if err.(*Error).StatusCode() != http.StatusBadRequest {
+		t.Fatalf("TestStatusCode failed, %d", err.(*Error).StatusCode())
+	}
+
+	if err.(*Error).InternalError() {
+		t.Fatalf("TestInternalError failed")
+	}
+
+	err = NewErrorf(ErrInvalidParams, "")
+	if len(err.Error()) == 0 {
+		t.Fatalf("TestNewErrorf with empty detial failed")
+	}
+}
+
+func TestRegisterErrors(t *testing.T) {
+	RegisterErrors(map[int32]string{503999: "test1", 1: "none"})
+
+	e := NewError(503999, "test2")
+	if e.Message != "test1" {
+		t.Fatalf("TestRegisterErrors failed")
+	}
 }
